+++ CHANGELOG.md
@@ -4,9 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.0.15]
+### Added
+- server:rename -- renames a server
+- BaseCommand -- cleaner code
+- Prompts for server name will now have auto-complete
+### Changed
+- Refactored all commands -- to use BaseCommand
+### Fixed
+- Some typos
 ### Removed
-- #6 illuminate/filesystem from composer.json
+- illuminate/filesystem from composer.json [#6](https://github.com/TNovalis/FiveM-ServerManager/pull/6)
 
 ## [0.0.14] - 2017-11-23
 ### Changed